Rename ingestion to extract.
